+++ packager/app/protection_system_flags.cc
@@ -12,5 +12,5 @@ DEFINE_string(additional_protection_systems,
               "",
               "Generate additional protection systems in addition to the "
               "native protection system provided by the key source. Supported "
-              "protection systems include Widevine, PlayReady, and "
+              "protection systems include Widevine, PlayReady, FairPlay, and "
               "CommonSystem (https://goo.gl/s8RIhr).");